Rename a few components, for clarity
