+++ src/app/pages/ParatimeDashboardPage/RosePriceCard.tsx
