+++ src/app/pages/ParatimeDashboardPage/SnapshotCard.tsx
