+++ src/app/pages/TransactionDetailPage/index.tsx
@@ -25,7 +25,7 @@ import { BlockLink } from '../../components/Blocks/BlockLink'
 import { TransactionLink } from '../../components/Transactions/TransactionLink'
 import { TransactionLogs } from '../../components/Transactions/Logs'
 import { useRequiredScopeParam } from '../../hooks/useScopeParam'
-import { DashboardLink } from '../DashboardPage/DashboardLink'
+import { DashboardLink } from '../ParatimeDashboardPage/DashboardLink'
 import { getNameForTicker, getTickerForNetwork, Ticker } from '../../../types/ticker'
 import { TokenPriceInfo, useTokenPrice } from '../../../coin-gecko/api'
 import { CurrentFiatValue } from './CurrentFiatValue'
